Add the CloudflareTunnel API (#39)
* Create CloudflareTunnel API

* Flesh out cloudflare tunnel type

* First go at the api token

* Remove the CRD API token

* Test boilerplate and rename types

* Install mockgen and create a client abstraction

* Run go generate in the generate target

* Move the mock around

* WIP on impl

* MVP

* Some more tests

* Refactor

* Move the client package

* Update client interface

* Existing tunnel tests

* Tests with existing tunnel id

* Failure mode tests

* Setup for the happy path

* Finalizer tests

* Add the finalizer in the one flow thats working

* Remove unnecessary tests and fix setup

* Adds tunnel update logic

Extends the controller to update existing Cloudflare tunnels.

The tunnel reconciliation now checks if the tunnel already
exists based on the tunnel ID stored in the status. If the tunnel
exists, it fetches the tunnel details from Cloudflare and updates the
status. If not, it creates a new tunnel.

Adds error handling for the tunnel retrieval.

* Defines CloudflareTunnelHealth type

Introduces a dedicated `CloudflareTunnelHealth` type to represent the health status of a Cloudflare Tunnel.

This change replaces the generic string type previously used for the status field with a custom enum. This provides better type safety and validation for the tunnel's health status. It also includes kubebuilder validation to ensure that only valid status values are used.

* Removes unused ListTunnels function

Removes the ListTunnels function and related code from the client and mocks.
This function is no longer needed, simplifying the client interface and reducing unnecessary complexity.

* Adds tunnel deletion functionality

Implements the ability to delete tunnels. This includes adding the DeleteTunnel method to the Client interface and its concrete implementation.
Also includes adding the DeleteTunnel method to the MockClient for testing purposes.

* Adds TODO comment for time format

Adds a TODO comment to remind developers to switch to metav1.Time for representing timestamps.

* Adds deletion tests for CloudflareTunnel

Adds tests to verify the CloudflareTunnel controller
correctly handles deletion, including successful and
failed Cloudflare API calls.

* Fixes CloudflareTunnel deletion issues

Ensures the CloudflareTunnel controller properly handles resource deletion by:

- Ignoring "not found" errors during resource retrieval in the cleanup process.
- Explicitly deleting the resource after finalizers are added in deletion tests.
- Matching against any DeleteTunnel call and propagating Cloudflare errors.
- Checking for resource deletion by verifying "not found" errors after reconciliation.

These changes improve the reliability and correctness of the CloudflareTunnel controller's deletion logic.

* Adds Cloudflare Tunnel deletion

Implements deletion of Cloudflare Tunnels via finalizers.

This change adds a finalizer to CloudflareTunnel resources and a
deletion function that removes the tunnel from Cloudflare's API when
the resource is deleted in Kubernetes. It also removes the finalizer
once the tunnel is successfully deleted, or if the tunnel ID is empty,
indicating that the tunnel was never successfully created.

Fixes a bug where resources would be stuck deleting.

* Lint fixes

* Regen dist