Tunnel e2e test (#43)
* Setup stuff for using the api token in e2e tests

* Recommend direnv

* Adds Cloudflare client to reconciler

The Cloudflare tunnel reconciler now uses a dedicated
Cloudflare client to interact with the Cloudflare API.

This provides better separation of concerns and testability.

* Configures sample CloudflareTunnel spec

Populates sample CloudflareTunnel spec with account ID,
config source, and name.

Removes the 'Name' column from the CloudflareTunnel status
as it's redundant.

Removes deletion timestamp logging.

* Regen dist

* Adds e2e test for Cloudflare Tunnel

Adds an end-to-end test case that validates the creation, verification, and deletion of a Cloudflare Tunnel resource.
This ensures the operator correctly manages Cloudflare Tunnels.
It also configures the necessary environment variable for the Cloudflare Account ID.