Handle Cloudflared Spec modifications (#52)
* Updates DaemonSet on template change

Ensures that the DaemonSet is updated when the pod template specification in the Cloudflared resource is modified.

This change addresses the need to propagate template updates to the managed DaemonSet, ensuring that configuration changes are reflected in the running pods.

* Tests DaemonSet and Deployment updates

Adds integration tests to verify that updates to the pod template spec in the Cloudflared custom resource propagate correctly to the managed DaemonSet and Deployment.

These tests ensure the controller reconciles changes to the container specifications as expected.

* Updates existing app resources

Adds functionality to update existing application
resources, specifically Deployments and DaemonSets.
This ensures that changes to the Cloudflared resource
are propagated to the underlying application.