Support changes to the cloudflared kind (#54)
* Adds cloudflared kind to status

Adds the cloudflared kind to the status of the Cloudflared resource.

This allows users to easily identify the type of application
managing the cloudflared instance.

* Adds Cloudflared status update tests

Adds tests to verify the Cloudflared status is updated
correctly based on the deployed Kind.

Includes tests to ensure that when the Kind is changed from
Deployment to DaemonSet, the Deployment is deleted, and a
DaemonSet is created.

* Creates app only if status kind is empty

Ensures the app is created only when the
Cloudflared resource's status kind is empty.

This prevents unnecessary app creation attempts and
improves reconciliation logic.
It also populates the status kind.

* Adds tests for Kind changes

Adds tests to verify the controller's behavior
when the Cloudflared resource Kind is updated.

These tests ensure the controller correctly deletes
the old resource and creates the new one.

* Handles Kind changes in Cloudflared spec

Ensures that the controller recreates the app if the
Kind specified in the Cloudflared spec differs from the
observed Kind of the existing app. This allows users to
seamlessly switch between Deployment and DaemonSet without
manual intervention.
Also fixes a typo in a log message.