Add tunnel reference values to the Cloudflared API (#57)
* Add tunnel reference values to the Cloudflared API

* Adds tunnel token retrieval

Implements the ability to retrieve a tunnel token. This enables enhanced authentication and authorization mechanisms for accessing and managing tunnels.

* Adds account ID to Cloudflared configuration

Introduces `CloudflaredConfigInline` to group tunnel ID and account ID.

This allows specifying the account ID alongside the tunnel ID
directly in the Cloudflared configuration.

The change also updates the deepcopy logic for the new struct.

* Adds inline config support for Cloudflared

Enables users to specify Cloudflared configuration directly within the resource definition, offering a more streamlined configuration process.

Retrieves tunnel token using the provided account ID and tunnel ID.

* Adds Cloudflare mock to controller tests

The Cloudflare mock is injected into the CloudflaredReconciler
in the controller tests. This allows for more thorough and
isolated testing of the controller logic.

* Fetches Cloudflare tunnel token

This commit enhances the cloudflared controller to fetch the tunnel token from Cloudflare's API.

It introduces a lookup mechanism using the Tunnel ID and Account ID provided in the Cloudflared resource spec.
The token is then injected as an environment variable into the cloudflared pod.
This allows cloudflared instances to authenticate with Cloudflare without requiring a configuration file.

* Adds inline config support for Cloudflared

Implements the ability to configure Cloudflared resources using inline configurations.
This allows users to define tunnel ID and account ID directly within the resource spec.

Additionally, ensures the controller does not attempt to retrieve the tunnel token from Cloudflare
when the CLOUDFLARE_API_TOKEN environment variable is not set.

* Implement the thing

* Regen dist

* Fix typo

* Remove the redundant function