Add an owner reference between Cloudflared and CloudflareTunnel to ensure proper deletion ordering (#59)
* Wip on tunnel owner ref

* Fixes Cloudflared owner reference test

Updates the Cloudflared controller test to correctly assert
that owner references are not controller references. This
ensures the correct behavior of the controller regarding
ownership and deletion of resources.

* Adds owner reference to CloudflareTunnel

Ensures proper deletion order by adding an owner reference from the CloudflareTunnel to the Cloudflared resource.

This ensures that all cloudflared connections are closed before deleting the tunnel.

* Lint

* Also lint