Clean up some redundant logic to be more controller-like (#64)
* Makes CloudflaredStatus.Kind optional

Updates the CloudflaredStatus.Kind field to be a pointer to
CloudflaredKind.

This allows the Kind field to be nil, indicating that the kind
of the app managing the cloudflared instance is not known.

* Handles app recreation when kind changes

Addresses an issue where the application was not properly recreated when the `Kind` field in the specification was modified.

Updates the reconciliation logic to detect changes in the `Kind` field. If a change is detected, the existing app is deleted, and a new one is created with the updated `Kind`.

This ensures that the application is always in sync with the desired state specified in the Cloudflared resource.

* Fixes Cloudflared Kind status update issue.

Addresses an issue where the Cloudflared Kind status was not being properly cleared during updates.
When the Kind of a Cloudflared resource changes (e.g., from DaemonSet to Deployment), the controller
now correctly clears the Kind status field to reflect the change. This ensures the controller properly handles resource transitions and avoids potential conflicts.
Additionally, reconciliation is now requeued after the Kind is changed.

* Updates Cloudflared reconciliation logic

Simplifies the Cloudflared reconciliation logic by removing tunnel reference checks.
This change also improves the handling of owned application resources by adding a status condition when the resource is not found, and requeues reconciliation to attempt creation.