Support creating cloudflared instances via template on CloudflareTunnel (#75)
* Add cloudflared template to the tunnel API

* Refactors CloudflareTunnel controller tests

Updates the tests for the CloudflareTunnel controller to use
`ContainElements` and `SatisfyAll` matchers. This simplifies the test
assertions and improves readability. Also adds a test to ensure
that the CloudflareTunnel resource is marked as progressing and not
available during reconciliation.

Removes the deprecated `meta.FindStatusCondition` function.

* Adds tunnel status conditions

Adds `Degraded` and `Progressing` status conditions to the CloudflareTunnel resource, providing more granular feedback during reconciliation.

Updates tunnel creation and deletion to set appropriate status conditions.

Applies tunnel ID and account ID config to Cloudflared resources.

* Adds cloudflared template test

Adds a test case for when the cloudflared template
is configured, verifying a Cloudflared resource
is created matching the specification, and updates
the tunnel status accordingly.

* Adds Cloudflared selector test case

Adds a test case to verify that a Cloudflared resource is not created
when the CloudflareTunnel selector does not match existing Cloudflared
resources.

Also, uses namespaced name directly in get call.

* Applies config to new Cloudflared instances

Creates a new Cloudflared instance from the CloudflareTunnel template if no matching Cloudflared instances are found.

This ensures that configurations are applied to newly created Cloudflared instances based on the provided template and label selector.

* Preserve unkown fields apparently

* No need to log the template

* Improves CloudflareTunnel deletion handling

Refactors the CloudflareTunnel controller to handle various
scenarios during deletion. It now correctly removes the finalizer
when the tunnel exists, is already marked as deleted, or does not
exist in the Cloudflare API. It also ensures that any owned
Cloudflared resources are deleted.

* Deletes cloudflared resource

Ensures that the controller deletes the cloudflared resource when its selector matches.
This prevents orphaned resources.

* Adds owner reference checks to tests

Adds checks to ensure that the controller sets the expected owner references on managed resources.

This ensures proper garbage collection and ownership of resources created by the controller.

* Ensures cloudflared is properly owned

Ensures that the cloudflared resource is correctly owned by the CloudflareTunnel resource.

This establishes the proper parent/child relationship, which
is important for cascading deletes and ensures that the cloudflared
resource is cleaned up when the tunnel is deleted.

* Refactors controller reconciliation in tests

Moves the instantiation of the controller reconciler to before the actual reconciliation process within the test. This ensures the reconciler is initialized with the latest state and dependencies before each reconciliation attempt in the test scenario.

* Adds test to preserve finalizer on delete failure

Adds a test case to ensure that the finalizer is not
removed from a CloudflareTunnel resource when the delete
operation fails. This ensures proper cleanup can occur
even when initial deletion attempts encounter errors.

Removes a duplicate test case.

* Deletes owned Cloudflared instances

Ensures that when a CloudflareTunnel resource is deleted, any
Cloudflared instances it owns are also deleted. This prevents
orphaned resources and ensures proper cleanup.

* Regen dist